Rename Util File, Add Email to PR (#847)
