+++ qiskit/optimization/util.py
