+++ test/optimization/test_optimization_problem_to_negative_value_oracle.py
@@ -16,7 +16,7 @@
 
 from test.optimization import QiskitOptimizationTestCase
 from qiskit.optimization.converters import OptimizationProblemToNegativeValueOracle
-from qiskit.optimization.portfolio_util import get_qubo_solutions
+from qiskit.optimization.util import get_qubo_solutions
 from qiskit import QuantumCircuit, Aer, execute
 import numpy as np
